+++ aquaterm/AQTClientManager.h
@@ -20,7 +20,7 @@
    id _activePlotKey;
    //void (*_errorHandler)(NSString *msg);	/* A callback function optionally installed by the client */
    //void (*_eventHandler)(long index, NSString *event); /* A callback function optionally installed by the client */
-   void (^_eventBlock)(NSInteger index, NSString *event);
+   void (^_eventBlock)(int index, NSString *event);
    void (^_errorBlock)(NSString *msg);
    id _eventBuffer;
    int32_t _logLimit;
@@ -34,9 +34,9 @@
 - (void)terminateConnection;
 - (void)setActivePlotKey:(id)newActivePlotKey;
 @property (copy) void (^errorBlock)(NSString *msg);
-@property (copy) void (^eventBlock)(NSInteger index, NSString *event);
+@property (copy) void (^eventBlock)(int index, NSString *event);
 - (void)setErrorHandler:(void (*)(NSString *errMsg))fPtr;
-- (void)setEventHandler:(void (*)(NSInteger index, NSString *event))fPtr;
+- (void)setEventHandler:(void (*)(int index, NSString *event))fPtr;
 
 - (void)logMessage:(NSString *)msg logLevel:(int32_t)level;
